# tools: port js2c.py to C++