--- a/deps/simdutf/simdutf.gyp
+++ b/deps/simdutf/simdutf.gyp
@@ -7,6 +7,7 @@
   'targets': [
     {
       'target_name': 'simdutf',
+      'toolsets': ['host', 'target'],
       'type': 'static_library',
       'include_dirs': ['.'],
       'direct_dependent_settings': {