… test/tools/test_js2c.py